# openapi/spec.yaml
...
openapi: "3.1.0"
info:
  title: JSON AI Studio API
  description: |-
    AI-powered configuration management platform — enable developers, product
    managers, and solution architects to create, modify, validate, and manage
    JSON configurations through natural language conversation.
  version: "0.1.0"
...
servers:
   - url: http://localhost:8000
     description: Local development server
...
paths:
...
   /api/sessions/{sessionId}/versions:
    get:
      tags: [sessions]
      summary: List all versions for a session
      description: |-
        Return all version snapshots (original, working, approved) with their
        timestamps. Used to populate the version sidebar in the frontend.
      operationId: listVersions
      security:
         - simpleApiKey: []
      parameters:
         - name: sessionId
           in: path
           required: true
           description: UUID of the session
           schema:
             type: string
      responses:
         '200':
          description: Version list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items:
                       $ref: '#/components/schemas/VersionSnapshot'
                required: [versions]
              example:
                versions:
                   - id: "v1"
                     parent_id: null
                     json_data: {}
                     label: "original"
                     created_at: "2026-06-21T00:00:00Z"
         '401':
           $ref: '#/components/responses/Error401'
         '404':
           $ref: '#/components/responses/Error404'
         '429':
           $ref: '#/components/responses/Error429'

    post:
      tags: [sessions]
      summary: Create a version snapshot
      description: |-
        Take a named snapshot of the current working JSON. Each snapshot is an
        immutable point-in-time state that can be restored later.
      operationId: createVersionSnapshot
      security:
         - simpleApiKey: []
      parameters:
         - name: sessionId
           in: path
           required: true
           description: UUID of the session
           schema:
             type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/CreateVersionRequest'
            example:
              label: "v1 — increase timeout"
              json_data:
                services:
                  api:
                    timeout: 60
      responses:
         '201':
          description: Version snapshot created
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/VersionSnapshot'
              example:
                id: "v2"
                parent_id: "v1"
                json_data: {}
                label: "v1 — increase timeout"
                created_at: "2026-06-21T00:01:00Z"
         '401':
           $ref: '#/components/responses/Error401'
         '404':
           $ref: '#/components/responses/Error404'
         '429':
           $ref: '#/components/responses/Error429'
   /api/sessions/{sessionId}/versions/select:
    post:
      tags: [sessions]
      summary: Select a version snapshot as the working baseline
      description: |-
        Load a version snapshot into the session: both working_json and
        baseline_json are reset to the snapshot's json_data (deep copies)
        and the conversation history is cleared.
      operationId: selectVersion
      security:
         - simpleApiKey: []
      parameters:
         - name: sessionId
           in: path
           required: true
           description: UUID of the session
           schema:
             type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/SelectVersionRequest'
            example:
              versionId: "v2"
      responses:
         '200':
          description: Version selected; session state reset to the snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  working_json:
                    type: object
                    additionalProperties: true
                  baseline_json:
                    type: object
                    additionalProperties: true
                  active_version_id:
                    type: string
                    nullable: true
                  versions:
                    type: array
                    items:
                       $ref: '#/components/schemas/VersionSnapshot'
                  conversation_history:
                    type: array
                    items:
                       $ref: '#/components/schemas/ChatTurn'
                required: [working_json, baseline_json, versions]
         '401':
           $ref: '#/components/responses/Error401'
         '404':
           $ref: '#/components/responses/Error404'
         '429':
           $ref: '#/components/responses/Error429'
   /api/sessions/{sessionId}/versions/restore:
    post:
      tags: [sessions]
      summary: Bulk-restore version snapshots into an empty session
      description: |-
        Restore previously saved version snapshots (e.g. from the browser's
        IndexedDB cache) into a freshly created session after a backend
        restart. Snapshot ids, labels, parent chains, and timestamps are
        preserved verbatim. Only allowed when the session has no versions
        yet — otherwise responds 409 so client retries are safe.
      operationId: restoreVersions
      security:
         - simpleApiKey: []
      parameters:
         - name: sessionId
           in: path
           required: true
           description: UUID of the session to restore into
           schema:
             type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/RestoreVersionsRequest'
            example:
              versions:
                 - id: "v1"
                   parent_id: null
                   json_data: {"services": {"api": {"timeout": 30}}}
                   label: "original"
                   created_at: "2026-07-01T00:00:00Z"
              working_json: {"services": {"api": {"timeout": 60}}}
              active_version_id: "v1"
      responses:
         '200':
          description: Versions restored; full session state returned
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/SessionResponse'
         '400':
           $ref: '#/components/responses/Error400'
         '401':
           $ref: '#/components/responses/Error401'
         '404':
           $ref: '#/components/responses/Error404'
         '409':
           $ref: '#/components/responses/Error409'
         '429':
           $ref: '#/components/responses/Error429'
   /api/sessions/{sessionId}/diffs/{diffId}/accept:
    post:
      tags: [sessions]
      summary: Apply an approved diff to the working JSON
      description: |-
        Merge the specified diff into the session's working_json. The diff is
        marked as accepted and the updated state becomes the new baseline for
        subsequent chat turns. An automatic validation run is triggered after
        acceptance; if validation fails, a warning toast is shown to the user
        but the diff remains applied.
      operationId: acceptDiff
      security:
         - simpleApiKey: []
      parameters:
         - name: sessionId
           in: path
           required: true
           description: UUID of the session
           schema:
             type: string
         - name: diffId
           in: path
           required: true
           description: UUID of the diff to accept
           schema:
             type: string
      responses:
         '200':
          description: Diff applied successfully
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/DiffAcceptResponse'
              example:
                success: true
                working_json: {}
                validation_passed: true
         '401':
           $ref: '#/components/responses/Error401'
         '404':
           $ref: '#/components/responses/Error404'
         '429':
           $ref: '#/components/responses/Error429'
   /api/sessions/{sessionId}/diffs/{diffId}/reject:
    post:
      tags: [sessions]
      summary: Revert a rejected diff from the working JSON
      description: |-
        Remove the specified diff from the working JSON. The diff is marked as
        rejected and does not affect the baseline state.
      operationId: rejectDiff
      security:
         - simpleApiKey: []
      parameters:
         - name: sessionId
           in: path
           required: true
           description: UUID of the session
           schema:
             type: string
         - name: diffId
           in: path
           required: true
           description: UUID of the diff to reject
           schema:
             type: string
      responses:
         '200':
          description: Diff rejected successfully
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/DiffRejectResponse'
              example:
                success: true
                working_json: {}
         '401':
           $ref: '#/components/responses/Error401'
         '404':
           $ref: '#/components/responses/Error404'
         '429':
           $ref: '#/components/responses/Error429'

   # =====================================================================
   # JSON UPLOAD & PARSING
   # =====================================================================
   /api/json/upload:
    post:
      tags: [upload]
      summary: Upload or paste JSON for parsing
      description: |-
        Accept a .json file upload or raw JSON body. Parse the input, validate
        syntax, generate a schema summary (top-level keys, nested depth, array
        lengths), and return the structured tree along with session metadata.
        The response includes a sessionId that the frontend uses for all
        subsequent chat operations.
      operationId: uploadJson
      security:
         - simpleApiKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: .json file to upload (max 20 MB)
                json_body:
                  type: string
                  description: Raw JSON string to parse as an alternative to file upload
              oneOf:
                 - required: [file]
                 - required: [json_body]
          application/json:
            schema:
              type: object
              properties:
                json_body:
                  type: string
                  description: Raw JSON string to parse
            example:
              json_body: '{"services": {"api": {"timeout": 30}}}'
      responses:
         '200':
          description: JSON parsed successfully
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/UploadResponse'
              example:
                session_id: "abc123"
                parsed_json:
                  services:
                    api:
                      timeout: 30
                schema_summary:
                  top_level_keys: ["services"]
                  nested_depth: 3
                  array_lengths: {}
         '400':
           $ref: '#/components/responses/Error400'
         '401':
           $ref: '#/components/responses/Error401'
         '413':
          description: Payload too large — file exceeds 20 MB limit
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/Error'
              example:
                error: "File size exceeds 20 MB limit"
                hint: "Please reduce the file size and try again."
         '429':
           $ref: '#/components/responses/Error429'

   # =====================================================================
   # CHAT (SSE STREAMING)
   # =====================================================================
...
   /api/diff:
    post:
      tags: [diff]
      summary: Compute diff between two JSON documents
      description: |-
        Return a structured field-level diff between old and new JSON objects.
        Supports nested objects, arrays, and scalar types. Used by the frontend
        to render the before/after side-by-side comparison view with color-coded
        diff indicators (green for added, orange for modified, red for deleted).
      operationId: computeDiff
      security:
         - simpleApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_json:
                  type: object
                  additionalProperties: true
                  description: The original JSON state (before)
                new_json:
                  type: object
                  additionalProperties: true
                  description: The modified JSON state (after)
              required: [old_json, new_json]
            examples:
              modify_value:
                summary: Modify a single scalar value
                value:
                  old_json:
                    services:
                      api:
                        timeout: 30
                  new_json:
                    services:
                      api:
                        timeout: 60
              add_field:
                summary: Add a new field
                value:
                  old_json:
                    services:
                      api: {}
                  new_json:
                    services:
                      api:
                        timeout: 30
              delete_field:
                summary: Remove an existing field
                value:
                  old_json:
                    services:
                      api:
                        timeout: 30
                  new_json:
                    services:
                      api: {}
      responses:
         '200':
          description: Structured diff between two JSON documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  diffs:
                    type: array
                    items:
                       $ref: '#/components/schemas/DiffEntry'
                    description: List of field-level changes
                required: [diffs]
              example:
                diffs:
                   - path: "/services/api/timeout"
                     operation: "modify"
                     old_value: 30
                     new_value: 60
         '400':
           $ref: '#/components/responses/Error400'
         '401':
           $ref: '#/components/responses/Error401'
         '429':
           $ref: '#/components/responses/Error429'

   # =====================================================================
   # EXPORT
   # =====================================================================
...
components:
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          description: UUID of the session
        name:
          type: string
          description: Human-readable session name
        working_json:
          type: object
          additionalProperties: true
          description: Current mutable state (arbitrary user-provided schema)
        versions:
          type: array
          items:
             $ref: '#/components/schemas/VersionSnapshot'
          description: Historical version snapshots
        conversation_history:
          type: array
          items:
             $ref: '#/components/schemas/ChatTurn'
          description: Full chat history for this session
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of session creation
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of last update
      required: [id, working_json, versions, conversation_history]
    VersionSnapshot:
      type: object
      properties:
        id:
          type: string
          description: UUID of this version snapshot
        parent_id:
          type: string
          nullable: true
          description: Reference to the predecessor version
        json_data:
          type: object
          additionalProperties: true
          description: The JSON state at time of snapshot
        label:
          type: string
          description: Human-readable label (e.g., "original", "v1 — increase timeout")
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of creation
      required: [id, json_data, label, created_at]
    ChatTurn:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
          description: Speaker of the chat turn
        content:
          type: string
          description: The message text
        diffs:
          type: array
          items:
             $ref: '#/components/schemas/DiffEntry'
          nullable: true
          description: Attached diff entries for assistant turns
      required: [role, content]

     # -- Wire Contracts --
    DiffEntry:
      type: object
      properties:
        path:
          type: string
          description: JSON Pointer path to the changed field (e.g., "/services/api/timeout")
        operation:
          type: string
          enum: [add, modify, delete]
          description: Type of change applied
        old_value:
          nullable: true
          description: Value before the change; null for "add" operations
        new_value:
          nullable: true
          description: Value after the change; null for "delete" operations
      required: [path, operation]
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        hint:
          type: string
          nullable: true
          description: Suggestion for how to resolve the error; null if not applicable
      required: [error]

     # -- Request/Response Wrappers --
...
    SessionResponse:
      allOf:
         - $ref: '#/components/schemas/Session'
      description: Full session state returned by GET /api/sessions/{id} and POST /api/sessions
    CreateVersionRequest:
      type: object
      properties:
        label:
          type: string
          description: Human-readable version label
        json_data:
          type: object
          additionalProperties: true
          description: JSON state to snapshot
      required: [label, json_data]
    SelectVersionRequest:
      type: object
      properties:
        versionId:
          type: string
          description: ID of the version snapshot to load
      required: [versionId]
    RestoreVersionsRequest:
      type: object
      properties:
        versions:
          type: array
          items:
             $ref: '#/components/schemas/VersionSnapshot'
          maxItems: 100
          description: Snapshots to restore verbatim (ids/labels/timestamps preserved)
        working_json:
          type: object
          additionalProperties: true
          nullable: true
          description: Working JSON to restore (unsaved work from the browser cache)
        baseline_json:
          type: object
          additionalProperties: true
          nullable: true
          description: Baseline JSON to restore; defaults to working_json when omitted
        active_version_id:
          type: string
          nullable: true
          description: Previously active version id; ignored if not among restored versions
      required: [versions]
...
    UploadResponse:
      type: object
      properties:
        session_id:
          type: string
          description: UUID of the created or retrieved session
        parsed_json:
          type: object
          additionalProperties: true
          description: The parsed JSON tree from the uploaded file
        schema_summary:
           $ref: '#/components/schemas/SchemaSummary'
      required: [session_id, parsed_json]
    SchemaSummary:
      type: object
      properties:
        top_level_keys:
          type: array
          items:
            type: string
          description: Top-level keys in the JSON structure
        nested_depth:
          type: integer
          minimum: 0
          description: Maximum nesting depth of the JSON
        array_lengths:
          type: object
          additionalProperties:
            type: integer
          description: 'Array lengths keyed by path (e.g., "/services": 3)'
      required: [top_level_keys, nested_depth]
...
    DiffAcceptResponse:
      type: object
      properties:
        success:
          type: boolean
        working_json:
          type: object
          additionalProperties: true
        validation_passed:
          type: boolean
          description: Whether the post-acceptance validation passed
      required: [success, working_json]
    DiffRejectResponse:
      type: object
      properties:
        success:
          type: boolean
        working_json:
          type: object
          additionalProperties: true
      required: [success, working_json]

     # -- HTTP Error Response Templates --
  responses:
    Error400:
      description: Bad request — invalid input format or missing required fields
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request body"
            hint: "Check your JSON syntax and required fields."
    Error401:
      description: Unauthorized — missing or invalid API key in X-API-Key header
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/Error'
          example:
            error: "Missing or invalid API key. Send valid X-API-Key header."
            hint: null
    Error404:
      description: Not found — session does not exist
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/Error'
          example:
            error: "Session not found"
            hint: "Check the session ID and try again."
    Error409:
      description: Conflict — session already contains version snapshots
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/Error'
          example:
            error: "Session already has versions"
            hint: "Restore is only allowed into a session with no versions."
    Error429:
      description: Too many requests — rate limit exceeded (10 requests per 60s window per API key)
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/Error'
          example:
            error: "Rate limited. 10 requests per 60s window."
            hint: null
...